requirements.txt
click==8.1.8
colorama==0.4.6
Flask==2.2.5
gunicorn==23.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
